Add dialects for supported databases for easier to use
