+++ dialect_postgres.go
@@ -1,14 +1,10 @@
 package gorm
 
 import (
-	"database/sql"
-	"database/sql/driver"
 	"fmt"
 	"reflect"
 	"strings"
 	"time"
-
-	"github.com/lib/pq/hstore"
 )
 
 type postgres struct {
@@ -55,7 +51,7 @@ func (postgres) DataTypeOf(field *StructField) string {
 				sqlType = "timestamp with time zone"
 			}
 		case reflect.Map:
-			if dataValue.Type() == hstoreType {
+			if dataValue.Type().Name() == "Hstore" {
 				sqlType = "hstore"
 			}
 		default:
@@ -106,51 +102,6 @@ func (s postgres) LastInsertIdReturningSuffix(tableName, key string) string {
 
 func (postgres) SupportLastInsertId() bool {
 	return false
-}
-
-var hstoreType = reflect.TypeOf(Hstore{})
-
-type Hstore map[string]*string
-
-func (h Hstore) Value() (driver.Value, error) {
-	hstore := hstore.Hstore{Map: map[string]sql.NullString{}}
-	if len(h) == 0 {
-		return nil, nil
-	}
-
-	for key, value := range h {
-		var s sql.NullString
-		if value != nil {
-			s.String = *value
-			s.Valid = true
-		}
-		hstore.Map[key] = s
-	}
-	return hstore.Value()
-}
-
-func (h *Hstore) Scan(value interface{}) error {
-	hstore := hstore.Hstore{}
-
-	if err := hstore.Scan(value); err != nil {
-		return err
-	}
-
-	if len(hstore.Map) == 0 {
-		return nil
-	}
-
-	*h = Hstore{}
-	for k := range hstore.Map {
-		if hstore.Map[k].Valid {
-			s := hstore.Map[k].String
-			(*h)[k] = &s
-		} else {
-			(*h)[k] = nil
-		}
-	}
-
-	return nil
 }
 
 func isByteArrayOrSlice(value reflect.Value) bool {
