+++ dialects/mysql/mysql.go
@@ -1,0 +1,3 @@
+package mysql
+
+import _ "github.com/go-sql-driver/mysql"
